made separate internal component
